requirements.txt
CherryPy>=3.2.2
Jinja2>=2.6
PyYAML>=3.10
SQLAlchemy>=0.7.5
Sphinx>=1.1.2
redis>=2.4.11
webassets>=0.6
python-dateutil==1.5
cdecimal>=2.3
MarkupSafe>=0.15

coverage>=3.5.1
nose>=1.1.2
nose-testconfig>=0.8
openstack.nose-plugin>=0.5
lazr.smtptest>=1.3
tox>=1.3

Sphinx>=1.1.2